#340 refactoring, passage des services de ResultView vers FacetsList
